changed controller file for qrCodeAuth
